requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==1.0.1
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.18.0
six==1.16.0
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
